// interfaces/Coin.ts
export interface Coin{
    btcPrice: string
    change: string
    coinrankingUrl: string
    color: string
    iconUrl: string
    listedAt: number
    lowVolume: boolean
    marketCap: string
    name: string
    price: string
    rank: number
    sparkLine: string[]
    symbol: string
    tier: number
    uuid: string
}